PC4/Entrega_3/requirements.txt
gradio
transformers
python-doctr[torch]
clean-text
pyspellchecker
nltk
scikit-learn
networkx
PyMuPDF
huggingface_hub
tqdm
rouge-score
natsort
torch
pandas
bert-extractive-summarizer
sentencepiece
